Core: finetuning allocator
git-svn-id: svn://ultimatepp.org/upp/trunk@13449 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/heaputil.cpp
@@ -18,7 +18,7 @@ void OutOfMemoryPanic(size_t size)
 }
 
 size_t Heap::huge_4KB_count;
-size_t Heap::free_4KB;
+int    Heap::free_4KB;
 size_t Heap::big_size;
 size_t Heap::big_count;
 size_t Heap::sys_size;
